+++ app/controllers/api/v1/room_settings_controller.rb
@@ -2,7 +2,7 @@
 
 module Api
   module V1
-    class RoomSettingsController < ApplicationController
+    class RoomSettingsController < ApiController
       skip_before_action :verify_authenticity_token # TODO: amir - Revisit this.
       before_action :find_room, only: %i[show update]
