+++ app/controllers/api/v1/rooms_controller.rb
@@ -2,7 +2,7 @@
 
 module Api
   module V1
-    class RoomsController < ApplicationController
+    class RoomsController < ApiController
       skip_before_action :verify_authenticity_token # TODO: amir - Revisit this.
       before_action :find_room, only: %i[show start recordings join status]
